backend/requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
langgraph
requests
python-multipart
PyPDF2
python-docx
pydantic
pydantic-settings
python-dotenv
realtimetts[kokoro]
torch
torchaudio
numpy
scipy
sse-starlette
nest-asyncio
tenacity
prometheus-client
websockets
